fix #11 close all streams while unzipping attester
Many builds randomly fail while trying to rename the temporary folder into
the target one. This may be due to some files remaining open inside the
directory.

Both streams are now closed as recommended by Apache Commons: calling
close inside the try block to receive errors in the catch block, and
calling IOUtils.closeQuietly in the finally block to ensure it is actually
closed if other I/O error interrupted the flow prior to that.

Input and output streams are also created using Apache Commons to avoid
having to create the intermediate streams.